refactor: S3Util infra 모듈로 이동
