+++ cherrypic-api/src/main/java/org/cherrypic/domain/image/dto/request/AlbumImageUploadRequest.java
@@ -8,8 +8,8 @@ import jakarta.validation.constraints.NotNull;
 import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
-import org.cherrypic.domain.image.enums.FileExtension;
 import org.cherrypic.global.annotation.Enum;
+import org.cherrypic.s3.enums.FileExtension;
 
 public record AlbumImageUploadRequest(
         @NotEmpty(message = "업로드할 피일들의 정보는 비워둘 수 없습니다.") @Valid @Schema(description = "업로드 요청 리스트")
