+++ cherrypic-infrastructure/build.gradle
@@ -13,6 +13,7 @@ repositories {
 }
 
 dependencies {
+	implementation project(':cherrypic-common')
 	implementation 'org.springframework.boot:spring-boot-starter-data-redis'
 
 	implementation 'io.awspring.cloud:spring-cloud-starter-aws:2.4.4'
